+++ backend/cwa_lib/pages/su_manage_vdbs.py
@@ -50,7 +50,8 @@ su_manage_vdbs__table_options = TableOptions(
     create__ask_columns=gvdbs_create_columns,
     update__ask_columns=gvdbs_edit_columns,
     delete__ask_columns=['gvdbs_id', ] + gvdbs_edit_columns,
-    order_by__allow=gvdbs_order_columns
+    order_by__allow=gvdbs_order_columns,
+    filter__allow=['group_id', 'enabled'],
 )
 
 class SuManageVDBsTableRead(GenericTableRead):