+++ frontend/src/tables/elements/TableHeader.tsx
@@ -2,6 +2,7 @@ import type { createTableStore } from "../TableStoreFactory";
 import CreateRowButton from "./CreateRowButton";
 import ExportButton from "./ExportButton";
 import HeaderColumnsRow from "./HeaderColumnsRow";
+import HeaderFilterRow from "./HeaderFilterRow";
 import PaginationControls from "./PaginationControls/PaginationControls";
 import TableReloadButton from "./TableReloadButton";
 import TableTitle from "./TableTitle";
@@ -25,6 +26,7 @@ export default function TableHeader({ useStore }: Props) {
         </th>
       </tr>
       <HeaderColumnsRow useStore={useStore} />
+      <HeaderFilterRow useStore={useStore} />
     </thead>
   );
 }